# [js/webgpu] Optimize transpose