@qjia7 2024-09-04
> Can we add testcases that exercise the newly added code if not already exists?

Done. Please take another look, thanks.